fix(core): detached `ViewRef`s are destroyed when `ViewContainerRef` is destroyed
This commit implements in the `ViewContainerRef` instances the behaviors
of keeping track of the `ViewRef` of components or embedded views it has
detached, so that when the `ViewContainerRef` instance itself gets destroy
those detached views (dynamically created by this `ViewContainerRef`)
are also be destroyed (and thus, run their own OnDestroy hooks). Earlier,
no detached `ViewRef` was ever destroyed if the user didn't do it manually.
Therefore, if the `ViewContainerRef` was destroyed, detached `ViewRef`s
weren't able to clear their subscription `Observable`s and asychronous
processes like `setTimeout` or `setInterval`, in the end you will have
memory leaks.